backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
bcrypt==4.2.0
PyJWT==2.9.0
gunicorn==22.0.0
